Move trickyarithmetic to a separate file
